Rename BasePageView.js to basePageView.js (#1431)
I just enabled the example extension on my mac and everything was peachy. Then I pushed it up to my dev linux server and the capitalization of this file caused an issue for me. Thought I'd submit it in case anyone else ran into similar issues in a case-sensitive environment.