Fix issues with the form state when navigating between collection settings pages (#20949)
* Code optimization in collection fields component

* Add changeset

* Improve subpath check in edit route guard

Previously, it would also match paths at the same level and thus wrongly
allowing to exit page. For example with two collections having the same base name
(example, example2).

* Account for undefined collection prop

* Switch back to `useItem` but re-initialize

This page/component is a bit of a special case, because it's re-used with different
props (collection) without being unmounted (and thus not re-initialized) when
navigating via provided links on relational fields or history.

We still want to rely on `useItem` to fetch the latest state of a
collection incl. translation as well as having the usual vars & methods
available like `loading`, `save`, ... which we can use with v-form.

However in this case, we need to completely re-initialize `useItem` per
collection, not just fetching the new collection on page change.

* Update changeset

* Full reset in useItems on refresh

* Don't clear edits

Should be checked & cleared externally, e.g. via edit guard

* Missed to remove `ìmmediate` option

---------

Co-authored-by: Brainslug <br41nslug@users.noreply.github.com>