fix(cache): enable expiration for permissions namespace in system cache (#26295)
Enables the expiration of keys for the "permissions" namespace in the Redis
system cache. Updates the @directus/memory package to pass TTL when adding new
keys. It adds the Time-to-Live (TTL) field to the CacheConfigRedis
interface. Updates the KvRedis.set() method to use this TTL value for setting
the expiration for keys. The TTL is sourced from the CACHE_SYSTEM_TTL
environment variable, defaulting to -1 (no expiration) if not specified.

Refs: #25513

Co-authored-by: Rijk van Zanten <rijkvanzanten@me.com>